+++ changelog.md
@@ -7,3 +7,10 @@
 
 -   Add `use_headings` and `slug` options for MkDocs backend.
 -   Fix inappropriate translation of image URLs into local paths in MkDocs preprocessor.
+
+
+# 1.0.3
+
+-   Add logs.
+-   Update for Foliant 1.0.4: Pass logger to spinner.
+-   Require Foliant 1.0.4.